Add support for AWS logs endpoints
Initial support is for streaming of logs. This is not an exact
match but close enough. Users will want a way of accessing the logs
programmatically.

Overloads `logStreamName` to mean the service name. Log endpoint
events are not returned by default by can be retrieved by using
`logStreamName` with `logs`.